split the submit form
